Move duplicate presenter.
